+++ doc/whats_new.rst
@@ -10,6 +10,9 @@ Enhancements
       a Nifti image. This avoids the step of unmasking the attribute
       `components_` which is true for older versions.
 
+    - New object :class:`nilearn.regions.Parcellations` for learning brain
+      parcellations on fmri data.
+
 Changes
 -------
 
@@ -186,9 +189,6 @@ Enhancements
 
     - Reworking and renaming of connectivity measures example. Renamed from
       plot_connectivity_measures to plot_group_level_connectivity.
-
-    - New object :class:`nilearn.regions.Parcellations` for learning brain
-      parcellations on fmri data.
 
     - Tighter bounding boxes when using add_contours for plotting.